+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/Task.swift
@@ -11,24 +11,3 @@ protocol Task {
     func process()
 }
 
-import Foundation
-
-struct Loan: Task {
-    let name: String = "대출업무"
-    
-    private let duration: Double = 1.1
-    
-    func process() {
-        Thread.sleep(forTimeInterval: self.duration)
-    }
-}
-
-struct Deposit: Task {
-    let name: String = "예금업무"
-    
-    private let duration: Double = 0.7
-    
-    func process() {
-        Thread.sleep(forTimeInterval: self.duration)
-    }
-}
